mtl/ofi: break from progress loop when events are read
Once any number of events are read, return immediately, rather than
waiting for fi_cq_read() to return FI_EAGAIN or an error. This can
improve observed latency if the user application is in a blocking call
waiting for us to return. Breaking here also means
ofi_progress_event_count serves as an upper bound for the total number
of events read in a single call (without it, we could read far more, as
long as new events continue to arrive).

Signed-off-by: Eric Badger <eric@badgerio.us>